Fix bugs in code , add builting handeling and fixed bugs in it

+++ built_ins/ft_exit.c
@@ -6,7 +6,7 @@
 /*   By: mbousset <mbousset@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/11 08:41:04 by salahian          #+#    #+#             */
-/*   Updated: 2025/04/24 15:14:09 by mbousset         ###   ########.fr       */
+/*   Updated: 2025/04/24 22:01:20 by mbousset         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,8 +31,7 @@ long	ft_atoi(char *str, int *flag)
 		return (*flag = 1, 0);
 	while (str[i] >= '0' && str[i] <= '9')
 	{
-		if ((sign == 1 && re > (LONG_MAX - (str[i] - '0')) / 10) || (sign == -1
-				&& re > ((unsigned long)LONG_MAX + 1 - (str[i] - '0')) / 10))
+		if ((sign == 1 && re > (LONG_MAX - (str[i] - '0')) / 10) || (sign == -1 && (unsigned long)re > ((unsigned long)LONG_MAX + 1 - (str[i] - '0')) / 10))
 			return (*flag = 1, 0);
 		re = (re * 10) + (str[i] - '0');
 		i++;
